requirements.txt
oandapyV20
python-dotenv
supabase
pandas